TorrentList: split context menu actions
